Fix SRC_DATA struct for 64-bit Linux (LP64 data model)
libsamplerate uses C 'long' for frame counts in SRC_DATA. On 64-bit
Linux (LP64), C 'long' is 8 bytes, but the original code used 'int'
(4 bytes), causing struct layout mismatch that would corrupt frame
counts or crash the native call.

Changed frame count fields from 'int' to 'nint' (native int) which
correctly maps to 8 bytes on 64-bit and 4 bytes on 32-bit platforms.

+++ src/MultiRoomAudio/Audio/LibSampleRate/AdaptiveSampleRateConverter.cs
@@ -210,11 +210,12 @@ public sealed class AdaptiveSampleRateConverter : IDisposable
         }
 
         // Copy output from pinned buffer
-        var outputSamples = srcData.OutputFramesGen * _channels;
+        // Cast from nint to int - frame counts will never exceed int range for audio
+        var outputSamples = (int)srcData.OutputFramesGen * _channels;
         _pinnedOutputBuffer.AsSpan(0, outputSamples).CopyTo(output);
 
-        inputFramesUsed = srcData.InputFramesUsed;
-        return srcData.OutputFramesGen;
+        inputFramesUsed = (int)srcData.InputFramesUsed;
+        return (int)srcData.OutputFramesGen;
     }
 
     /// <summary>
